fix folder name typo
